# misc: Compile regexps only once