It is unnecessary to compile the same regexps over and over with every
call to `filterFrames`.

This improves the performance of `NewStacktrace()` reducing both CPU time
and memory allocation.

```
$ go1.11.13 test -run=- -count=5 -bench=. -benchmem | tee old.txt
$ # make changes ...
$ go1.11.13 test -run=- -count=5 -bench=. -benchmem | tee new.txt
$ benchstat old.txt new.txt
name             old time/op    new time/op    delta
NewStacktrace-8    54.0µs ± 0%    12.6µs ± 1%  -76.73%  (p=0.008 n=5+5)

name             old alloc/op   new alloc/op   delta
NewStacktrace-8    92.1kB ± 0%     7.3kB ± 0%  -92.08%  (p=0.008 n=5+5)

name             old allocs/op  new allocs/op  delta
NewStacktrace-8      92.0 ± 0%      15.0 ± 0%  -83.70%  (p=0.008 n=5+5)
```